requirements.txt
fastapi
uvicorn
opencv-python
numpy
pytest
httpx
jinja2
requests
line-bot-sdk
